Fix isSubstringHelper method
Added a check to ensure that the code wasn't defaulting to true for cases where the substring fell at the end of the string but the part of the string where the substring started was not long enough to contain the entire substring.

Modified a pertinent test to focus on the broken part of the code (isSubstringHelper) by shortening the text string.